+++ workspace/solutions/rust/src/s0/s9/mod.rs
@@ -1,1 +1,2 @@
+pub mod s0;
 pub mod s1;