+++ workspace/solutions/rust/src/s0/s9/s0/s7/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0907_00;
+pub mod s_0907_test;
+
+pub type TSolution = fn(Vec<i32>) -> i32;